+++ esp32_arm/src/AGV/mover_driver/single_wheel_driver/single_wheel_driver_base.h
@@ -2,7 +2,7 @@
 
 class SingleWheelDriverBase{
     public:
-        virtual void SetVelocity_in_percent(int8_t velocity);
+        virtual void SetVelocity_in_percent(int velocity);
         
     private:
