+++ esp32_arm/src/AGV/mover_driver/single_wheel_driver/single_wheel_h_bridge_pwm_driver.h
@@ -7,7 +7,7 @@ class SingleWheel_HBridgePwmDriver: public SingleWheelDriverBase{
     public:
         SingleWheel_HBridgePwmDriver(uint8_t pin_dir, uint8_t pin_pwm);
         SingleWheel_HBridgePwmDriver(uint8_t pin_pwm, Adafruit_MCP23X17* mcp_23018_dir_only, uint8_t pin_dir);
-        void SetVelocity_in_percent(int8_t velocity) override;
+        void SetVelocity_in_percent(int velocity) override;
 
 
     private:
